moved DiffParameter to own module
